Refactor: Removed unused filed and moved files from HCDP to SensorOVerview
