--- webstack/libs/applications/src/lib/apps/HCDP/data/hcdpStationData.tsx
@@ -1,58 +1,0 @@
-export const hcdpStationData = [
-  {
-    uuid: '479252997392297490-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-07-17T16:00:34.959-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '254.12',
-      name: 'Upper Kahikinui 0145',
-      observer: 'HIMesonet',
-      network: 'HIMesonet',
-      island: 'MA',
-      elevation_m: '1034',
-      lat: '20.644215',
-      lng: '-156.284703',
-    },
-    created: '2023-07-17T16:00:34.959-05:00',
-    _links: {
-      self: { href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/479252997392297490-242ac1110-0001-012' },
-      permissions: { href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/479252997392297490-242ac1110-0001-012/pems' },
-      owner: { href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj' },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '6699009887094697490-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-07-17T15:58:10.144-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '92.11',
-      name: 'Keaau 0202',
-      observer: 'HIMesonet',
-      network: 'HIMesonet',
-      island: 'HA',
-      elevation_m: '182',
-      lat: '19.6061748',
-      lng: '-155.0515231',
-    },
-    created: '2023-07-17T15:58:10.144-05:00',
-    _links: {
-      self: { href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/6699009887094697490-242ac11f-0001-012' },
-      permissions: { href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/6699009887094697490-242ac11f-0001-012/pems' },
-      owner: { href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj' },
-      associationIds: [],
-    },
-  }
-];
