--- webstack/libs/applications/src/lib/apps/HCDP/data/hcdpStationData_legacy.tsx
@@ -1,3423 +1,0 @@
-export const hcdpStationData = [
-  {
-    uuid: '3989160470763344366-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-06-15T20:00:41.941-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '598.2',
-      name: 'Honolimaloo 0412',
-      observer: 'HIMesonet',
-      network: 'HIMesonet',
-      island: 'MO',
-      elevation_m: '402',
-      lat: '21.131411',
-      lon: '-156.758626',
-      nws_id: '031HI',
-    },
-    created: '2023-06-15T20:00:41.941-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3989160470763344366-242ac11f-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3989160470763344366-242ac11f-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '3981816076687184366-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-06-15T20:00:41.770-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '598.1',
-      name: 'Keopukaloa 0411',
-      observer: 'HIMesonet',
-      network: 'HIMesonet',
-      island: 'MO',
-      elevation_m: '187',
-      lat: '21.145283',
-      lon: '-156.729459',
-      nws_id: '030HI',
-    },
-    created: '2023-06-15T20:00:41.770-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3981816076687184366-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3981816076687184366-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2482460718877175315-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-05-24T16:13:44.338-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '191.41',
-      name: 'Lalamilo 0252',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'BI',
-      elevation_m: '802',
-      lat: '20.019528',
-      lon: '-155.677085',
-      nws_id: '027HI',
-    },
-    created: '2023-05-24T16:13:44.338-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2482460718877175315-242ac11f-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2482460718877175315-242ac11f-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '3296620912664056301-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-05-24T16:08:49.396-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '68.61',
-      name: 'Keahuolu 0241',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'MA',
-      elevation_m: '537',
-      lat: '19.668669',
-      lon: '-155.957474',
-      nws_id: '029HI',
-    },
-    created: '2023-05-24T16:08:49.396-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3296620912664056301-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3296620912664056301-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '1189144293139017235-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-04-06T16:49:41.512-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1116.2',
-      name: 'Lower Limahuli 0603',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'KA',
-      elevation_m: '31',
-      lat: '22.219805',
-      lon: '-159.575195',
-      nws_id: '024HI',
-    },
-    created: '2023-04-06T16:49:41.512-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/1189144293139017235-242ac11f-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/1189144293139017235-242ac11f-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '547256421369254381-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-04-06T16:43:12.445-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '181.3',
-      name: 'Kehena 0251',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'BI',
-      elevation_m: '1158',
-      lat: '20.122834',
-      lon: '-155.749329',
-      nws_id: '025HI',
-    },
-    created: '2023-04-06T16:43:12.445-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/547256421369254381-242ac11f-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/547256421369254381-242ac11f-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '3574362359478161901-242ac11f-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-30T15:26:47.066-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '283',
-      name: 'Hamoa 0165',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'MA',
-      elevation_m: '106',
-      lat: '20.719497',
-      lon: '-156.002356',
-      nws_id: '023HI',
-    },
-    created: '2023-03-30T15:26:47.066-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3574362359478161901-242ac11f-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/3574362359478161901-242ac11f-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '6142446169540455955-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-06T21:10:51.932-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '254.11',
-      name: 'Kahikikinui 0144',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'MA',
-      elevation_m: '713',
-      lat: '20.633945',
-      lon: '-156.273889',
-      nws_id: '019HI',
-    },
-    created: '2023-03-06T21:10:51.932-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/6142446169540455955-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/6142446169540455955-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '7354915437201255955-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-06T21:10:23.702-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '141.1',
-      name: 'Kulaimano 0212',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'BI',
-      elevation_m: '260',
-      lat: '19.834341',
-      lon: '-155.122435',
-      nws_id: '021HI',
-    },
-    created: '2023-03-06T21:10:23.702-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/7354915437201255955-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/7354915437201255955-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '7615748801087335955-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-06T21:10:17.629-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '126.2',
-      name: 'Kanakaleonui 0211',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'BI',
-      elevation_m: '2352',
-      lat: '19.845036',
-      lon: '-155.362586',
-      nws_id: '022HI',
-    },
-    created: '2023-03-06T21:10:17.629-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/7615748801087335955-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/7615748801087335955-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '8454117421342855661-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-06T21:09:22.287-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '12.16',
-      name: 'Kaiholena 0231',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'BI',
-      elevation_m: '631',
-      lat: '19.16877',
-      lon: '-155.57035',
-      nws_id: '018HI',
-    },
-    created: '2023-03-06T21:09:22.287-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/8454117421342855661-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/8454117421342855661-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '9108324359696805395-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:55:04.180-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1132.11',
-      name: 'Common Ground 0602',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'KA',
-      elevation_m: '112',
-      lat: '22.197439',
-      lon: '-159.42123',
-      nws_id: '015HI',
-    },
-    created: '2023-03-02T14:55:04.180-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/9108324359696805395-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/9108324359696805395-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '9219134515933605395-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:55:01.600-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.11',
-      name: 'Waipa 0601',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'KA',
-      elevation_m: '5',
-      lat: '22.202647',
-      lon: '-159.518833',
-      nws_id: '014HI',
-    },
-    created: '2023-03-02T14:55:01.600-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/9219134515933605395-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/9219134515933605395-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '5607424573650506221-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:53:37.311-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '785.12',
-      name: 'Lyon 0501',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'OA',
-      elevation_m: '151',
-      lat: '21.333',
-      lon: '-157.8025',
-      nws_id: '011HI',
-    },
-    created: '2023-03-02T14:53:37.311-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/5607424573650506221-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/5607424573650506221-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '5419176157066826221-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:53:32.928-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '775.11',
-      name: 'Nuuanu Res1 0502',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'OA',
-      elevation_m: '117',
-      lat: '21.339124',
-      lon: '-157.836859',
-      nws_id: '012HI',
-    },
-    created: '2023-03-02T14:53:32.928-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/5419176157066826221-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/5419176157066826221-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2278223623829066221-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:52:19.797-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '434.11',
-      name: 'Piiholo 0115',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'MA',
-      elevation_m: '637',
-      lat: '20.8415',
-      lon: '-156.2948',
-      nws_id: '017HI',
-    },
-    created: '2023-03-02T14:52:19.797-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2278223623829066221-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2278223623829066221-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '816603303327306221-242ac1110-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:51:45.765-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '267.11',
-      name: 'Keokea 0116',
-      observer: 'HiMesonet',
-      network: 'HiMesonet',
-      island: 'MA',
-      elevation_m: '871',
-      lat: '20.7067',
-      lon: '-156.3554',
-      nws_id: '016HI',
-    },
-    created: '2023-03-02T14:51:45.765-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/816603303327306221-242ac1110-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/816603303327306221-242ac1110-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '657603614029386221-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:51:42.064-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '257.11',
-      name: 'NAKULA HN-143',
-      observer: 'HiMesonet',
-      network: 'Halenet',
-      island: 'MA',
-      elevation_m: '1632',
-      lat: '20.674726',
-      lon: '-156.233409',
-    },
-    created: '2023-03-02T14:51:42.064-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/657603614029386221-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/657603614029386221-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '22289708927479315-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2023-03-02T14:51:26.234-06:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '201.4',
-      name: 'Kamuela Airport',
-      observer: 'NWS',
-      network: 'NWS',
-      island: 'BI',
-      elevation_m: '813.658',
-      lat: '19.9984',
-      lon: '-155.6743',
-      nws_id: 'PHMU',
-    },
-    created: '2023-03-02T14:51:26.234-06:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/22289708927479315-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/22289708927479315-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2803244274067247596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.597-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1147',
-      name: 'Fld 960-Moloaa',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '79',
-      lat: '22.17444115',
-      lon: '-159.3360919',
-    },
-    created: '2022-03-17T22:40:27.597-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2803244274067247596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2803244274067247596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2798906357098287596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.496-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1146',
-      name: 'Moloaa Dairy',
-      observer: 'RAWS',
-      network: 'RAWS',
-      island: 'KA',
-      elevation_m: '85.344',
-      lat: '22.18333333',
-      lon: '-159.3375',
-      ncei_id: 'USR0000HMOL',
-      nws_id: 'MLDH1',
-      nesdis_id: 'AAB6D15A',
-    },
-    created: '2022-03-17T22:40:27.496-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2798906357098287596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2798906357098287596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2794353691764527596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.390-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1145',
-      name: 'MOLOAA 1145',
-      observer: 'LIHUE PLANTATION',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '101',
-      lat: '22.17972',
-      lon: '-159.33194',
-      ncei_id: 'USC00516529',
-    },
-    created: '2022-03-17T22:40:27.390-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2794353691764527596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2794353691764527596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2789371529701167596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.274-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1144.1',
-      name: 'ANAHOLA 2.8 N',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '74',
-      lat: '22.1869',
-      lon: '-159.3236',
-      ncei_id: 'US1HIKI0021',
-    },
-    created: '2022-03-17T22:40:27.274-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2789371529701167596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2789371529701167596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2784002820581167596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.149-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1144',
-      name: 'Field 7-A',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '61',
-      lat: '22.18607715',
-      lon: '-159.3188701',
-    },
-    created: '2022-03-17T22:40:27.149-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2784002820581167596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2784002820581167596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2778333463750447596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:27.017-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1143',
-      name: 'Field 30',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '70',
-      lat: '22.20218961',
-      lon: '-159.3472027',
-    },
-    created: '2022-03-17T22:40:27.017-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2778333463750447596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2778333463750447596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2773136553322287596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.896-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1142',
-      name: 'Camp 1',
-      observer: 'HAWN CANNERIES',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '84',
-      lat: '22.1777722',
-      lon: '-159.3388696',
-    },
-    created: '2022-03-17T22:40:26.896-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2773136553322287596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2773136553322287596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2768154391258927596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.780-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1141',
-      name: 'Field 36',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '111',
-      lat: '22.19443816',
-      lon: '-159.3672022',
-    },
-    created: '2022-03-17T22:40:26.780-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2768154391258927596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2768154391258927596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2763558776252207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.673-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1140',
-      name: 'Field 27 (300)',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '91',
-      lat: '22.20221107',
-      lon: '-159.3755354',
-    },
-    created: '2022-03-17T22:40:26.673-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2763558776252207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2763558776252207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2758190067132207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.547-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1137',
-      name: 'KOLOKO RESERVOIR 113',
-      observer: 'KILAUEA SUGAR',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '224',
-      lat: '22.18140774',
-      lon: '-159.3777576',
-      ncei_id: 'USC00514758',
-    },
-    created: '2022-03-17T22:40:26.547-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2758190067132207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2758190067132207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2754023948855087596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.451-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1136',
-      name: 'Field 23',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '107',
-      lat: '22.19944652',
-      lon: '-159.3888685',
-    },
-    created: '2022-03-17T22:40:26.451-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2754023948855087596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2754023948855087596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2750072578942767596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.359-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1135',
-      name: 'KILAUEA FIELD 17',
-      observer: 'KILAUEA SUGAR',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '128',
-      lat: '22.18502918',
-      lon: '-159.4005349',
-      ncei_id: 'USC00514566',
-    },
-    created: '2022-03-17T22:40:26.359-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2750072578942767596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2750072578942767596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2745863510992687596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.261-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1134.2',
-      name: 'KILAUEA 0.5 SE',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '79',
-      lat: '22.2057',
-      lon: '-159.4048',
-      ncei_id: 'US1HIKI0006',
-    },
-    created: '2022-03-17T22:40:26.261-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2745863510992687596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2745863510992687596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2741482644350767596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.159-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1134.1',
-      name: 'KILAUEA 0.3 SE',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '96',
-      lat: '22.208',
-      lon: '-159.4062',
-      ncei_id: 'US1HIKI0001',
-    },
-    created: '2022-03-17T22:40:26.159-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2741482644350767596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2741482644350767596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2736113935230767596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:26.034-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1134',
-      name: 'KILAUEA 1134',
-      observer: 'RAPOZO N',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '96',
-      lat: '22.21194093',
-      lon: '-159.4072015',
-      ncei_id: 'USC00514561',
-    },
-    created: '2022-03-17T22:40:26.034-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2736113935230767596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2736113935230767596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2730487528073007596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.903-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1133',
-      name: 'KILAUEA POINT 1133',
-      observer: 'USCG',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '55',
-      lat: '22.23135314',
-      lon: '-159.4013684',
-      ncei_id: 'USC00514568',
-    },
-    created: '2022-03-17T22:40:25.903-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2730487528073007596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2730487528073007596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2725677164701487596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.790-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1132.2',
-      name: 'Hwy Fld 41',
-      observer: 'PRINCEVILLE RANCH',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '43',
-      lat: '22.21473095',
-      lon: '-159.4333123',
-    },
-    created: '2022-03-17T22:40:25.790-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2725677164701487596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2725677164701487596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2721038600021807596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.683-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1132.1',
-      name: 'Kilauea Fld 3',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '99',
-      lat: '22.20168903',
-      lon: '-159.4255345',
-    },
-    created: '2022-03-17T22:40:25.683-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2721038600021807596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2721038600021807596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2715669890901807596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.558-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1132',
-      name: 'Kilauea Fld 3 #1',
-      observer: 'KILAUEA SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '99',
-      lat: '22.20779162',
-      lon: '-159.4255346',
-    },
-    created: '2022-03-17T22:40:25.558-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2715669890901807596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2715669890901807596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2710644779165487596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.441-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1131.7',
-      name: 'HanaleiRainGageatHanalei',
-      observer: 'USGS',
-      network: 'USGS',
-      island: 'KA',
-      elevation_m: '18.288',
-      lat: '22.17958333',
-      lon: '-159.4663889',
-    },
-    created: '2022-03-17T22:40:25.441-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2710644779165487596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2710644779165487596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2706049164158767596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.334-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1131.5',
-      name: 'PRINCEVILLE 2.4 SE',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '119',
-      lat: '22.1996',
-      lon: '-159.458',
-      ncei_id: 'US1HIKI0007',
-    },
-    created: '2022-03-17T22:40:25.334-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2706049164158767596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2706049164158767596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2701195851114287596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.221-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1131.4',
-      name: 'Mauka Field 31',
-      observer: 'PRINCEVILLE RANC',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '134',
-      lat: '22.19587896',
-      lon: '-159.4527565',
-    },
-    created: '2022-03-17T22:40:25.221-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2701195851114287596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2701195851114287596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2695955991013167596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:25.099-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1131.1',
-      name: 'Kapaka',
-      observer: 'MCBRYDE SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '194',
-      lat: '22.17673899',
-      lon: '-159.4588675',
-    },
-    created: '2022-03-17T22:40:25.099-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2695955991013167596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2695955991013167596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2691231526987567596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.989-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1131',
-      name: 'KALIHWAI RESERVOIR',
-      observer: 'KILAUEA SUGAR',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '122',
-      lat: '22.18615576',
-      lon: '-159.4288679',
-      ncei_id: 'USC00512974',
-    },
-    created: '2022-03-17T22:40:24.989-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2691231526987567596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2691231526987567596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2687022459037487596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.891-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.8',
-      name: 'PRINCEVILLE 0.8 ESE',
-      observer: 'COOP',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '86.9',
-      lat: '22.217157',
-      lon: '-159.474314',
-      ncei_id: 'US1HIKI0028',
-    },
-    created: '2022-03-17T22:40:24.891-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2687022459037487596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2687022459037487596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2682641592395567596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.789-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.7',
-      name: 'PRINCEVILLE 0.5 E',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '57',
-      lat: '22.2242',
-      lon: '-159.4777',
-      ncei_id: 'US1HIKI0013',
-    },
-    created: '2022-03-17T22:40:24.789-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2682641592395567596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2682641592395567596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2677874178697007596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.678-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.6',
-      name: 'Lake 6',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '64',
-      lat: '22.22390641',
-      lon: '-159.4719231',
-    },
-    created: '2022-03-17T22:40:24.678-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2677874178697007596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2677874178697007596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2671818274809647596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.537-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.5',
-      name: 'Lake 2',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '76',
-      lat: '22.2200242',
-      lon: '-159.4741452',
-    },
-    created: '2022-03-17T22:40:24.537-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2671818274809647596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2671818274809647596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2665848270268207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.398-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.4',
-      name: 'Wood 9',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '67',
-      lat: '22.21891848',
-      lon: '-159.481923',
-    },
-    created: '2022-03-17T22:40:24.398-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2665848270268207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2665848270268207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2661166755915567596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.289-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.3',
-      name: 'Ocean 7',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '40',
-      lat: '22.22613387',
-      lon: '-159.4888675',
-    },
-    created: '2022-03-17T22:40:24.289-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2661166755915567596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2661166755915567596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2656528191235887596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.181-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.2',
-      name: 'Puu Pua',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '37',
-      lat: '22.22142279',
-      lon: '-159.4972007',
-    },
-    created: '2022-03-17T22:40:24.181-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2656528191235887596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2656528191235887596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2651202431788847596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:24.057-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118.1',
-      name: 'Kulaiki',
-      observer: 'PRINCEVILLE HANALEI',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '61',
-      lat: '22.22473699',
-      lon: '-159.4691453',
-    },
-    created: '2022-03-17T22:40:24.057-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2651202431788847596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2651202431788847596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2646263219398447596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.942-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1118',
-      name: 'Makai Fld 16',
-      observer: 'PRINCEVILLE RANCH',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '52',
-      lat: '22.22169532',
-      lon: '-159.4874786',
-    },
-    created: '2022-03-17T22:40:23.942-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2646263219398447596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2646263219398447596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2641366956681007596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.828-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.9',
-      name: 'PRINCEVILLE AIRPORT NEAR KILAUEA 4W',
-      observer: 'NWS',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '88.497',
-      lat: '22.2114',
-      lon: '-159.443',
-      nws_id: 'PRIH1',
-      nesdis_id: 'B15952F6',
-    },
-    created: '2022-03-17T22:40:23.828-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2641366956681007596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2641366956681007596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2636513643636527596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.715-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.8',
-      name: 'Hanalei',
-      observer: 'NWS',
-      network: 'HydroNet-UaNet',
-      island: 'KA',
-      elevation_m: '4.8768',
-      lat: '22.1982002',
-      lon: '-159.4951',
-      ncei_id: 'USC00513386',
-      nws_id: 'HNIH1',
-    },
-    created: '2022-03-17T22:40:23.715-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2636513643636527596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2636513643636527596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2631918028629807596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.608-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.7',
-      name: 'Shop',
-      observer: 'EAGLE CTY DEV CORP',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '67',
-      lat: '22.21364628',
-      lon: '-159.4788675',
-    },
-    created: '2022-03-17T22:40:23.608-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2631918028629807596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2631918028629807596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2626678168528687596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.486-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.5',
-      name: 'Hanalei',
-      observer: 'HAWN CANNERIES',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '76',
-      lat: '22.21224881',
-      lon: '-159.4588676',
-    },
-    created: '2022-03-17T22:40:23.486-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2626678168528687596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2626678168528687596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2622168452867887596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.381-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.4',
-      name: 'Fld P-24',
-      observer: 'HAWN CANNERIES',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '84',
-      lat: '22.21002604',
-      lon: '-159.452201',
-    },
-    created: '2022-03-17T22:40:23.381-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2622168452867887596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2622168452867887596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2617916435244847596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.282-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117.1',
-      name: 'Hanalei-Deverill',
-      observer: 'EGK DEVERILL',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '3',
-      lat: '22.20006472',
-      lon: '-159.5022006',
-    },
-    created: '2022-03-17T22:40:23.282-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2617916435244847596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2617916435244847596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2613234920892207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.173-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1117',
-      name: 'PRINCEVILLE RANCH 1117',
-      observer: 'PRINCEVILLE RANCH',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '90',
-      lat: '22.213196',
-      lon: '-159.4722008',
-      ncei_id: 'USC00518165',
-    },
-    created: '2022-03-17T22:40:23.173-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2613234920892207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2613234920892207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2607436715042607596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:23.038-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1116.1',
-      name: 'HANALEI 3.0 W, HI US',
-      observer: 'CoCoRaHs',
-      network: 'CoCoRaHs',
-      island: 'KA',
-      elevation_m: '10.457',
-      lat: '22.2103',
-      lon: '-159.5492',
-      ncei_id: 'US1HIKI0024',
-    },
-    created: '2022-03-17T22:40:23.038-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2607436715042607596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2607436715042607596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2603399445784367596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.944-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1115.1',
-      name: 'PH Wainiha',
-      observer: 'NWS',
-      network: 'HydroNet-UaNet',
-      island: 'KA',
-      elevation_m: '35.052',
-      lat: '22.1959991',
-      lon: '-159.5563',
-      nws_id: 'WNHH1',
-    },
-    created: '2022-03-17T22:40:22.944-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2603399445784367596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2603399445784367596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2599147428161327596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.845-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1115',
-      name: 'PH WAINIHA 1115',
-      observer: 'MCBRYDE SUGAR',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '35',
-      lat: '22.1959',
-      lon: '-159.5564',
-      ncei_id: 'USC00518155',
-    },
-    created: '2022-03-17T22:40:22.845-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2599147428161327596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2599147428161327596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2594036417079087596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.726-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1114',
-      name: 'ANAHOLA 1114',
-      observer: 'LIHUE PLANTATION',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '55',
-      lat: '22.13418691',
-      lon: '-159.3038709',
-      ncei_id: 'USC00510145',
-    },
-    created: '2022-03-17T22:40:22.726-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2594036417079087596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2594036417079087596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2589097204688687596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.611-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1113.1',
-      name: 'Field HC208',
-      observer: 'HAWN FRUIT PACKERS',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '30',
-      lat: '22.10672074',
-      lon: '-159.3055377',
-    },
-    created: '2022-03-17T22:40:22.611-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2589097204688687596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2589097204688687596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2583986193606447596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.492-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1113',
-      name: 'M9',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '46',
-      lat: '22.11005029',
-      lon: '-159.3055377',
-    },
-    created: '2022-03-17T22:40:22.492-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2583986193606447596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2583986193606447596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2578445685794607596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.363-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1112',
-      name: 'KEALIA 1112',
-      observer: 'LIHUE PLANTATION',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '3',
-      lat: '22.09950672',
-      lon: '-159.3055378',
-      ncei_id: 'USC00513982',
-    },
-    created: '2022-03-17T22:40:22.363-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2578445685794607596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2578445685794607596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2574236617844527596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.265-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1111.4',
-      name: 'Anahola',
-      observer: 'HAWN FRUIT PACKERS',
-      network: 'HydroNet-UaNet',
-      island: 'KA',
-      elevation_m: '75.2856',
-      lat: '22.1373997',
-      lon: '-159.312',
-      nws_id: 'ANHH1',
-      nesdis_id: '1560029C',
-    },
-    created: '2022-03-17T22:40:22.265-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2574236617844527596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2574236617844527596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2570113449240367596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.169-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1111.3',
-      name: 'HC-8',
-      observer: 'HAWN FRUIT PACKERS',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '49',
-      lat: '22.16832672',
-      lon: '-159.3222034',
-    },
-    created: '2022-03-17T22:40:22.169-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2570113449240367596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2570113449240367596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2564572941428527596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:22.040-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1111.2',
-      name: 'Mol 18',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '46',
-      lat: '22.16333171',
-      lon: '-159.3205368',
-    },
-    created: '2022-03-17T22:40:22.040-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2564572941428527596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2564572941428527596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2559504880019247596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.922-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1111',
-      name: 'Anahola',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '12',
-      lat: '22.14335296',
-      lon: '-159.3155371',
-    },
-    created: '2022-03-17T22:40:21.922-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2559504880019247596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2559504880019247596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2554179120572207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.798-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1110',
-      name: 'HALAULA 1110',
-      observer: 'LIHUE PLANTATION',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '78',
-      lat: '22.11505514',
-      lon: '-159.317204',
-      ncei_id: 'USC00510935',
-    },
-    created: '2022-03-17T22:40:21.798-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2554179120572207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2554179120572207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2549282857854767596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.684-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1107.3',
-      name: 'KAPAA 1.7 E',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '65',
-      lat: '22.0948',
-      lon: '-159.3278',
-      ncei_id: 'US1HIKI0008',
-    },
-    created: '2022-03-17T22:40:21.684-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2549282857854767596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2549282857854767596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2545589185980207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.598-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1107.2',
-      name: 'E Kauai Water 2',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '58',
-      lat: '22.09397345',
-      lon: '-159.3238707',
-    },
-    created: '2022-03-17T22:40:21.598-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2545589185980207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2545589185980207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2541208319338287596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.496-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1107.1',
-      name: 'Kapaa',
-      observer: 'STATE',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '15',
-      lat: '22.08840889',
-      lon: '-159.3072047',
-    },
-    created: '2022-03-17T22:40:21.496-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2541208319338287596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2541208319338287596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2536097308256047596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.377-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1107',
-      name: 'EKW 1 (120)',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '37',
-      lat: '22.08675145',
-      lon: '-159.3155377',
-    },
-    created: '2022-03-17T22:40:21.377-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2536097308256047596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2536097308256047596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2530427951425327596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.245-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1104.2',
-      name: 'M 19',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '24',
-      lat: '22.07261625',
-      lon: '-159.3369262',
-    },
-    created: '2022-03-17T22:40:21.245-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2530427951425327596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2530427951425327596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2524715644921647596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:21.112-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1104.1',
-      name: 'Sta 35 (Station 35)',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '24',
-      lat: '22.07844384',
-      lon: '-159.3372039',
-    },
-    created: '2022-03-17T22:40:21.112-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2524715644921647596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2524715644921647596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2518702690707247596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.972-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1104',
-      name: 'KAPAKA STABLES 1104',
-      observer: 'LIHUE PLANTATION',
-      network: 'COOP',
-      island: 'KA',
-      elevation_m: '53',
-      lat: '22.08177006',
-      lon: '-159.332204',
-      ncei_id: 'USC00513159',
-    },
-    created: '2022-03-17T22:40:20.972-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2518702690707247596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2518702690707247596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2513376931260207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.848-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1102.2',
-      name: 'Field 207-A',
-      observer: 'HAWN FRUIT PACKER',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '107',
-      lat: '22.12809836',
-      lon: '-159.320537',
-    },
-    created: '2022-03-17T22:40:20.848-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2513376931260207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2513376931260207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2509081963964207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.748-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1102.1',
-      name: 'Fld Lihue 3',
-      observer: 'HAWN CANNERIES',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '122',
-      lat: '22.1250533',
-      lon: '-159.3288702',
-    },
-    created: '2022-03-17T22:40:20.748-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2509081963964207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2509081963964207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2504786996668207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.648-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1102',
-      name: 'Field Makee 2b',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '187',
-      lat: '22.12340532',
-      lon: '-159.3505362',
-    },
-    created: '2022-03-17T22:40:20.648-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2504786996668207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2504786996668207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2500105482315567596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.539-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1101.1',
-      name: 'Spalding Monu',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '125',
-      lat: '22.12255786',
-      lon: '-159.3308146',
-    },
-    created: '2022-03-17T22:40:20.539-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2500105482315567596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2500105482315567596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2495853464692527596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.440-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1101',
-      name: 'Field Lihue 2',
-      observer: 'HAWN CANNERIES',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '122',
-      lat: '22.12310991',
-      lon: '-159.3272036',
-    },
-    created: '2022-03-17T22:40:20.440-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2495853464692527596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2495853464692527596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2490828352956207596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.323-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1100.1',
-      name: 'Kapahi (Mckee)',
-      observer: 'MCKEE SUGAR',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '93',
-      lat: '22.09843088',
-      lon: '-159.3472033',
-    },
-    created: '2022-03-17T22:40:20.323-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2490828352956207596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2490828352956207596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2485760291546927596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.205-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1100',
-      name: 'Kapahi',
-      observer: 'HAWN FRUIT PACKER',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '73',
-      lat: '22.09953145',
-      lon: '-159.3349814',
-    },
-    created: '2022-03-17T22:40:20.205-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2485760291546927596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2485760291546927596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2481078777194287596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.096-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1097.4',
-      name: 'KAPAA 1.9 WSW',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '157',
-      lat: '22.0829',
-      lon: '-159.3786',
-      ncei_id: 'US1HIKI0019',
-    },
-    created: '2022-03-17T22:40:20.096-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2481078777194287596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2481078777194287596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2476998558263087596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:20.001-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1097.3',
-      name: 'KAPAA 1.4 SW',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '111',
-      lat: '22.0805',
-      lon: '-159.3662',
-      ncei_id: 'US1HIKI0003',
-    },
-    created: '2022-03-17T22:40:20.001-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2476998558263087596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2476998558263087596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2471715748489007596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.878-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1097.2',
-      name: 'KAPAA 1.4 SSW',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '111',
-      lat: '22.0803',
-      lon: '-159.3659',
-      ncei_id: 'US1HIKI0012',
-    },
-    created: '2022-03-17T22:40:19.878-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2471715748489007596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2471715748489007596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2466690636752687596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.761-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1097.1',
-      name: 'Field 617',
-      observer: 'HAWN FRUIT PACKER',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '137',
-      lat: '22.08734482',
-      lon: '-159.3705361',
-    },
-    created: '2022-03-17T22:40:19.761-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2466690636752687596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2466690636752687596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2461665525016367596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.644-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1097',
-      name: 'EKW 4',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '102',
-      lat: '22.08178786',
-      lon: '-159.3588699',
-    },
-    created: '2022-03-17T22:40:19.644-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2461665525016367596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2461665525016367596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2455824369493807596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.508-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1095.1',
-      name: "Newlun's House",
-      observer: 'USGS',
-      network: 'USGS',
-      island: 'KA',
-      elevation_m: '198',
-      lat: '22.10677081',
-      lon: '-159.3722025',
-    },
-    created: '2022-03-17T22:40:19.508-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2455824369493807596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2455824369493807596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2450670408738607596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.388-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1095',
-      name: 'Mimino',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '85',
-      lat: '22.10841854',
-      lon: '-159.3455365',
-    },
-    created: '2022-03-17T22:40:19.388-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2450670408738607596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2450670408738607596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2445301699618607596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.263-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1094.3',
-      name: 'KAPAA 1.4 W',
-      observer: 'CoCoRaHS',
-      network: 'CoCoRaHS',
-      island: 'KA',
-      elevation_m: '139',
-      lat: '22.0936',
-      lon: '-159.3758',
-      ncei_id: 'US1HIKI0005',
-    },
-    created: '2022-03-17T22:40:19.263-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2445301699618607596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2445301699618607596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2439503493769007596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.127-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1094.2',
-      name: 'Kapahi',
-      observer: 'NWS',
-      network: 'HydroNet-UaNet',
-      island: 'KA',
-      elevation_m: '158.496',
-      lat: '22.1021004',
-      lon: '-159.3856',
-      ncei_id: 'USC00513171',
-      nws_id: 'KPIH1',
-      nesdis_id: '15608488',
-    },
-    created: '2022-03-17T22:40:19.127-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2439503493769007596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2439503493769007596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2434736080070447596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:19.017-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1094.1',
-      name: 'EKW3',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '107',
-      lat: '22.09011798',
-      lon: '-159.3672028',
-    },
-    created: '2022-03-17T22:40:19.017-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2434736080070447596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2434736080070447596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2429753918007087596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:18.901-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1094',
-      name: 'E Kaua W 3 (370)',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '113',
-      lat: '22.09844692',
-      lon: '-159.3735915',
-    },
-    created: '2022-03-17T22:40:18.901-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2429753918007087596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2429753918007087596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2424342259214127596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:18.775-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1092.4',
-      name: 'Field M1a',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '232',
-      lat: '22.13700848',
-      lon: '-159.3624803',
-    },
-    created: '2022-03-17T22:40:18.775-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2424342259214127596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2424342259214127596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2418672902383407596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:18.643-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1092.3',
-      name: 'Kaneha Res (800)',
-      observer: 'LIHUE PLANTATION',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '244',
-      lat: '22.12786006',
-      lon: '-159.373869',
-    },
-    created: '2022-03-17T22:40:18.643-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2418672902383407596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2418672902383407596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-  {
-    uuid: '2413132394571567596-242ac1111-0001-012',
-    owner: 'mcleanj',
-    schemaId: null,
-    internalUsername: null,
-    associationIds: [],
-    lastUpdated: '2022-03-17T22:40:18.514-05:00',
-    name: 'hcdp_station_metadata',
-    value: {
-      station_group: 'hawaii_climate_primary',
-      id_field: 'skn',
-      skn: '1092.2',
-      name: 'Fld W-6',
-      observer: 'HAWN FRUIT PACKERS',
-      network: 'STATE',
-      island: 'KA',
-      elevation_m: '137',
-      lat: '22.12674141',
-      lon: '-159.360536',
-    },
-    created: '2022-03-17T22:40:18.514-05:00',
-    _links: {
-      self: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2413132394571567596-242ac1111-0001-012',
-      },
-      permissions: {
-        href: 'https://agaveauth.its.hawaii.edu/meta/v2/data/2413132394571567596-242ac1111-0001-012/pems',
-      },
-      owner: {
-        href: 'https://agaveauth.its.hawaii.edu/profiles/v2/mcleanj',
-      },
-      associationIds: [],
-    },
-  },
-];
