+++ webstack/libs/applications/src/lib/apps/SensorOverview/SensorOverview.tsx
@@ -7,7 +7,7 @@
  */
 
 // Sage Imports
-import { useAppStore, useHexColor, useUIStore } from '@sage3/frontend';
+import { useAppStore } from '@sage3/frontend';
 import { App } from '../../schema';
 import { AppWindow } from '../../components';
 import { state as AppState } from './index';
@@ -37,28 +37,23 @@ import {
   Portal,
 } from '@chakra-ui/react';
 
-import { MdDelete, MdAdd, MdArrowDropDown, MdArrowDropUp, MdDoubleArrow } from 'react-icons/md';
+import { MdArrowDropDown, MdArrowDropUp, MdDoubleArrow } from 'react-icons/md';
 
 // Styling
 import './styling.css';
 import { useEffect, useRef, useState } from 'react';
 
 // Visualization imports
-import VariableCard from '../HCDP/viewers/VariableCard';
-import EChartsViewer from '../HCDP/viewers/EChartsViewer';
-import CurrentConditions from '../HCDP/viewers/CurrentConditions';
-import CustomizeWidgets, {
-  getFormattedDateTime1MonthBefore,
-  getFormattedDateTime1WeekBefore,
-  getFormattedDateTime1YearBefore,
-} from '../HCDP/menu/CustomizeWidgets';
-import StationMetadata from '../HCDP/viewers/StationMetadata';
-import FriendlyVariableCard from '../HCDP/viewers/FriendlyVariableCard';
-import StatisticCard from '../HCDP/viewers/StatisticCard';
-import MapViewer from '../HCDP/viewers/MapViewer';
+import VariableCard from './viewers/VariableCard';
+import EChartsViewer from './viewers/EChartsViewer';
+import CurrentConditions from './viewers/CurrentConditions';
+import { getFormattedDateTime1MonthBefore, getFormattedDateTime1WeekBefore, getFormattedDateTime1YearBefore } from './utils';
+import StationMetadata from './viewers/StationMetadata';
+import FriendlyVariableCard from './viewers/FriendlyVariableCard';
+import StatisticCard from './viewers/StatisticCard';
+import MapViewer from './viewers/MapViewer';
 
-import { checkAvailableVisualizations } from '../HCDP/menu/CustomizeWidgets';
-import LeafletWrapper from '../HCDP/LeafletWrapper';
+import { checkAvailableVisualizations } from './utils';
 import MapGL from './MapGL';
 
 export const getFormattedTimePeriod = (timePeriod: string) => {
@@ -95,19 +90,6 @@ function formatDuration(ms: number) {
   } else {
     return `less than a minute ago`;
   }
-}
-
-// Convert the dateTime to Chakra format
-function convertToChakraDateTime(dateTime: string) {
-  const year = dateTime.slice(0, 4);
-  const month = dateTime.slice(4, 6);
-  const day = dateTime.slice(6, 8);
-  const hours = dateTime.slice(8, 10);
-  const minutes = dateTime.slice(10, 12);
-
-  const formattedDateTime = `${year}-${month}-${day}T${hours}:${minutes}`;
-
-  return formattedDateTime;
 }
 
 function getFormattedDateTime24HoursBefore() {
@@ -168,8 +150,9 @@ function AppComponent(props: App): JSX.Element {
           new Date()
         )}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
       } else {
-        url = `https://api.mesowest.net/v2/stations/timeseries?STID=${String(s.stationNames)}&showemptystations=1&start=${props.data.state.widget.startDate
-          }&end=${convertToFormattedDateTime(new Date())}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
+        url = `https://api.mesowest.net/v2/stations/timeseries?STID=${String(s.stationNames)}&showemptystations=1&start=${
+          props.data.state.widget.startDate
+        }&end=${convertToFormattedDateTime(new Date())}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
       }
 
       const response = await fetch(url);
@@ -370,8 +353,8 @@ function AppComponent(props: App): JSX.Element {
                   />
                 ) : null}
                 {props.data.state.widget.visualizationType === 'line' ||
-                  props.data.state.widget.visualizationType === 'bar' ||
-                  props.data.state.widget.visualizationType === 'scatter' ? (
+                props.data.state.widget.visualizationType === 'bar' ||
+                props.data.state.widget.visualizationType === 'scatter' ? (
                   <EChartsViewer
                     stationNames={s.stationNames}
                     isLoaded={isLoaded}
@@ -438,16 +421,13 @@ function ToolbarComponent(props: App): JSX.Element {
   const createApp = useAppStore((state) => state.create);
 
   const updateState = useAppStore((state) => state.updateState);
-  const fitApps = useUIStore((state) => state.fitApps);
 
-  const [map, setMap] = useState<any>(null);
   const [stationMetadata, setStationMetadata] = useState([]);
   const [isLoaded, setIsLoaded] = useState<boolean>(false);
   const [isSortedOn, setIsSortedOn] = useState<{ name: string; direction: string }>({ name: 'NAME', direction: 'ascending' });
 
   // For color theme
   const textColor = useColorModeValue('gray.800', 'gray.50');
-  const drawerBackgroundColor: string = useColorModeValue('gray.50', 'gray.700');
   const headerBackgroundColor: string = useColorModeValue('white', 'gray.800');
   const accentColor: string = useColorModeValue('#DFDFDF', '#424242');
 
@@ -489,8 +469,9 @@ function ToolbarComponent(props: App): JSX.Element {
         new Date()
       )}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
     } else {
-      url = `https://api.mesowest.net/v2/stations/timeseries?STID=${String(s.stationNames)}&showemptystations=1&start=${props.data.state.widget.startDate
-        }&end=${convertToFormattedDateTime(new Date())}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
+      url = `https://api.mesowest.net/v2/stations/timeseries?STID=${String(s.stationNames)}&showemptystations=1&start=${
+        props.data.state.widget.startDate
+      }&end=${convertToFormattedDateTime(new Date())}&token=d8c6aee36a994f90857925cea26934be&complete=1&obtimezone=local`;
     }
 
     const response = await fetch(url);
@@ -572,31 +553,8 @@ function ToolbarComponent(props: App): JSX.Element {
     updateState(props._id, { widget: { ...s.widget, visualizationType: visualizationType } });
   };
 
-  const removeVisualizationsThatRequireMultipleStations = (availableVisualizations: { value: string; name: string }[]) => {
-    if (s.stationNames.length > 1) {
-      return availableVisualizations.filter((visualization) => {
-        return (
-          visualization.value !== 'variableCard' &&
-          visualization.value !== 'friendlyVariableCard' &&
-          visualization.value !== 'statisticCard'
-        );
-      });
-    } else {
-      return availableVisualizations;
-    }
-  };
-
-  const handleDateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const selectedDate = e.target.value;
-    const date = new Date(selectedDate);
-
-    const startDate = convertToFormattedDateTime(date);
-    updateState(props._id, { widget: { ...s.widget, startDate: startDate, timePeriod: 'custom' } });
-  };
-
   const handleSelectDateChange = (e: React.ChangeEvent<HTMLSelectElement>) => {
     const timePeriod = e.target.value;
-    const date = new Date();
     switch (timePeriod) {
       case 'previous24Hours':
         updateState(props._id, { widget: { ...s.widget, startDate: getFormattedDateTime24HoursBefore(), timePeriod: 'previous24Hours' } });
@@ -618,21 +576,6 @@ function ToolbarComponent(props: App): JSX.Element {
     }
   };
   const { isOpen, onOpen, onClose } = useDisclosure();
-
-  const handleAddSelectedStation = (station: { lat: number; lon: number; name: string; selected: boolean }) => {
-    const stationNames = s.stationNames;
-    stationNames.push(station.name);
-    updateState(props._id, { stationNames: stationNames });
-  };
-
-  const handleRemoveSelectedStation = (station: { lat: number; lon: number; name: string; selected: boolean }) => {
-    const stationNames = s.stationNames;
-    const index = stationNames.indexOf(station.name);
-    if (index > -1) {
-      stationNames.splice(index, 1);
-    }
-    updateState(props._id, { stationNames: stationNames });
-  };
 
   const handleFilterOn = (attributeName: string) => {
     const isFirstTime = isSortedOn.name !== attributeName || isSortedOn.direction === 'ascending';
@@ -776,25 +719,25 @@ function ToolbarComponent(props: App): JSX.Element {
                   {!isLoaded
                     ? null
                     : stationMetadata.map((station: any, index: number) => {
-                      const isSelected = s.stationNames.includes(station.STID);
-                      return (
-                        <tr key={index}>
-                          <td style={{ textAlign: 'center', width: '10px' }}>
-                            <Checkbox
-                              colorScheme="teal"
-                              isChecked={isSelected}
-                              onChange={(e) => handleChangeSelectedStation(e, station.STID)}
-                            />
-                          </td>
-                          <td>{station.NAME}</td>
-                          <td>{station.COUNTY}</td>
-                          <td style={{ textAlign: 'right' }}>{station.ELEVATION}</td>
-                          <td style={{ textAlign: 'right' }}>{Number(station.LATITUDE).toFixed(1)}</td>
-                          <td style={{ textAlign: 'right' }}>{Number(station.LONGITUDE).toFixed(1)}</td>
-                          {/* <td>variable</td> */}
-                        </tr>
-                      );
-                    })}
+                        const isSelected = s.stationNames.includes(station.STID);
+                        return (
+                          <tr key={index}>
+                            <td style={{ textAlign: 'center', width: '10px' }}>
+                              <Checkbox
+                                colorScheme="teal"
+                                isChecked={isSelected}
+                                onChange={(e) => handleChangeSelectedStation(e, station.STID)}
+                              />
+                            </td>
+                            <td>{station.NAME}</td>
+                            <td>{station.COUNTY}</td>
+                            <td style={{ textAlign: 'right' }}>{station.ELEVATION}</td>
+                            <td style={{ textAlign: 'right' }}>{Number(station.LATITUDE).toFixed(1)}</td>
+                            <td style={{ textAlign: 'right' }}>{Number(station.LONGITUDE).toFixed(1)}</td>
+                            {/* <td>variable</td> */}
+                          </tr>
+                        );
+                      })}
                 </table>
                 {!isLoaded ? (
                   <Box width="100%" height="100%" position="relative">
@@ -963,7 +906,7 @@ const GroupedToolbarComponent = (props: { apps: App[] }) => {
         mr="1rem"
         // value={widget.visualizationType}
         onChange={handleVisualizationChange}
-      // isDisabled={props.data.state.widget.yAxisNames[0] === 'Elevation, Longitude, Latitude, Name, Time'}
+        // isDisabled={props.data.state.widget.yAxisNames[0] === 'Elevation, Longitude, Latitude, Name, Time'}
       >
         {availableVisualizations().map((visualization: { value: string; name: string }, index: number) => {
           return (
