--- webstack/libs/applications/src/lib/apps/SensorOverview/data.tsx
@@ -1,62 +1,0 @@
-export const variableNames = [
-  {
-    variableName: 'date_time',
-  },
-  {
-    variableName: 'wind_cardinal_direction_set_1d',
-  },
-  {
-    variableName: 'wind_speed_set_1',
-  },
-  {
-    variableName: 'soil_temp_set_1',
-  },
-  {
-    variableName: 'soil_moisture_set_3',
-  },
-  {
-    variableName: 'soil_moisture_set_2',
-  },
-  {
-    variableName: 'outgoing_radiation_lw_set_1',
-  },
-  {
-    variableName: 'outgoing_radiation_sw_set_1',
-  },
-  {
-    variableName: 'dew_point_temperature_set_1d',
-  },
-  {
-    variableName: 'precip_accum_five_minute_set_1',
-  },
-  {
-    variableName: 'incoming_radiation_lw_set_1',
-  },
-  {
-    variableName: 'heat_index_set_1d',
-  },
-  {
-    variableName: 'net_radiation_lw_set_1',
-  },
-  {
-    variableName: 'net_radiation_sw_set_1',
-  },
-  {
-    variableName: 'net_radiation_set_1',
-  },
-  {
-    variableName: 'soil_moisture_set_1',
-  },
-  {
-    variableName: 'relative_humidity_set_1',
-  },
-  {
-    variableName: 'solar_radiation_set_1',
-  },
-  {
-    variableName: 'air_temp_set_1',
-  },
-  {
-    variableName: 'wind_direction_set_1',
-  },
-];
