+++ webstack/libs/applications/src/lib/apps/SensorOverview/data/stationData.ts
