+++ webstack/libs/applications/src/lib/apps/SensorOverview/types/types.tsx
@@ -1,0 +1,14 @@
+export type VariableProps = {
+  variableName: string;
+  stationName: string;
+  value: number;
+  average: number;
+  stdDev: number;
+  high: number;
+  low: number;
+  unit: string;
+  startDate: string;
+  endDate: string;
+  stationSTIDName: string;
+  color: string;
+};
