+++ webstack/libs/applications/src/lib/apps/SensorOverview/viewers/EChartsViewer.tsx
