+++ webstack/libs/applications/src/lib/apps/SensorOverview/viewers/FriendlyVariableCard.tsx
@@ -8,9 +8,7 @@
 
 import React, { useEffect, useState } from 'react';
 
-import { Box, Spinner, Text, Image, Divider, AbsoluteCenter, useColorMode, Icon } from '@chakra-ui/react';
-import VariableUnits from '../data/VariableUnits';
-import { stationColors, getColor } from '../../EChartsViewer/ChartManager';
+import { Box, Spinner, Text, Divider, useColorMode } from '@chakra-ui/react';
 
 import { App, AppState } from '@sage3/applications/schema';
 import variableUnits from '../data/VariableUnits';
@@ -85,11 +83,11 @@ type VariableProps = {
   startDate: string;
   endDate: string;
   stationSTIDName: string;
-  images: string[];
+
   color: string;
 };
 
-export default function VariableCard(
+export default function FriendlyVariableCard(
   props: {
     isLoaded: boolean;
     stationNames: string[];
@@ -129,12 +127,12 @@ export default function VariableCard(
         const sensorValues = props.stationMetadata[i].OBSERVATIONS[s.widget.yAxisNames[j]];
         if (sensorValues) {
           let unit = '';
-          let images: string[] = [];
+
           let color = '#ffffff';
-          for (let i = 0; i < VariableUnits.length; i++) {
-            if (s.widget.yAxisNames[j].includes(VariableUnits[i].variable)) {
-              unit = VariableUnits[i].unit;
-              images = VariableUnits[i].images;
+          for (let i = 0; i < variableUnits.length; i++) {
+            if (s.widget.yAxisNames[j].includes(variableUnits[i].variable)) {
+              unit = variableUnits[i].unit;
+
               color = variableUnits[i].color;
             }
           }
@@ -151,7 +149,7 @@ export default function VariableCard(
               stationSTIDName: props.stationMetadata[i].STID,
               startDate: props.stationMetadata[i].OBSERVATIONS['date_time'][0],
               endDate: props.stationMetadata[i].OBSERVATIONS['date_time'][props.stationMetadata[i].OBSERVATIONS['date_time'].length - 1],
-              images: images,
+
               color: color,
             });
           } else {
@@ -167,7 +165,7 @@ export default function VariableCard(
               stationSTIDName: props.stationMetadata[i].STID,
               startDate: props.startDate,
               endDate: '2022-04-25T19:55:00Z',
-              images: images,
+
               color: color,
             });
           }
@@ -271,7 +269,6 @@ export default function VariableCard(
                     startDate: props.startDate,
                     stationSTIDName: 'HI012',
                     endDate: '2022-04-25T19:55:00Z',
-                    images: [],
                   }
             }
           />
@@ -305,7 +302,6 @@ const Content = (props: {
     } else {
       setScaleToFontSize(props.size.height / Math.ceil(Math.sqrt(props.stationNames.length)) - 10);
     }
-    console.log(props.size.width, props.size.height);
   }, [JSON.stringify(props.size), JSON.stringify(props.stationNames)]);
   return (
     <>
@@ -341,18 +337,19 @@ const Content = (props: {
         display="flex"
         flexDirection="column"
         // justifyContent={'center'}
+        // flexWrap={'wrap'}
         alignContent="center"
         textAlign={'center'}
       >
         <Box bg="#2D62D2">
-          <Text textShadow={'black 2px 2px'} color="white" textAlign={'center'} fontSize={scaleToFontSize / 10}>
-            {props.variable.stationName}
+          <Text color="white" textShadow={'black 2px 2px'} fontSize={scaleToFontSize / 10}>
+            {variableName.join(' ')}
           </Text>
         </Box>
 
         <Box>
-          <Text mt={scaleToFontSize / 6} textShadow={'black 2px 2px'} fontSize={scaleToFontSize / 8}>
-            {variableName.join(' ')}
+          <Text mt={scaleToFontSize / 10} textAlign={'center'} fontSize={scaleToFontSize / 8}>
+            {props.variable.stationName}
           </Text>
         </Box>
 
@@ -393,6 +390,7 @@ const Content = (props: {
             fontSize={scaleToFontSize / 20}
             fontWeight="semibold"
             // lineHeight={'48px'}
+            mt={scaleToFontSize / 20}
           >
             <>{props.timeSinceLastUpdate}</>
           </Text>
@@ -402,6 +400,6 @@ const Content = (props: {
   );
 };
 
-VariableCard.defaultProps = {
+FriendlyVariableCard.defaultProps = {
   generateAllVariables: false,
 };
