+++ webstack/libs/applications/src/lib/apps/SensorOverview/viewers/MapViewer.tsx
@@ -23,7 +23,7 @@ import center from '@turf/center';
 import { useAppStore, useAssetStore, useUIStore, apiUrls } from '@sage3/frontend';
 import { Asset } from '@sage3/shared/types';
 import { App } from '../../../schema';
-import { state as AppState } from '../index';
+import { state as AppState } from '../../HCDP/index';
 // import { state as AppState } from './index';
 // import redMarker from './redMarker.png';
 
@@ -288,8 +288,9 @@ const MapViewer = (props: App & { isSelectingStations: boolean; isLoaded?: boole
             const el = document.createElement('div');
 
             el.innerHTML = `<div style="position: relative; ">
-            <div style=" border-radius: 50%; position: absolute; left: 50%; top: 50%; transform: scale(${s.stationScale
-              }); background-color: #2fa9ee; width: 20px; height: 20px; color: white; border: 2px solid black; display: flex; flex-direction: column; justify-content: center; ">
+            <div style=" border-radius: 50%; position: absolute; left: 50%; top: 50%; transform: scale(${
+              s.stationScale
+            }); background-color: #2fa9ee; width: 20px; height: 20px; color: white; border: 2px solid black; display: flex; flex-direction: column; justify-content: center; ">
             <p  style="font-size:5px; font-weight: bold; text-align: center">
             ${Number(latestValue).toFixed(1)}</p>
             </div>
