+++ webstack/libs/applications/src/lib/apps/SensorOverview/viewers/maplibre-gl.css
