+++ webstack/libs/applications/src/lib/apps/SensorOverview/viewers/redMarker.png
